first commit for a dv elite experiment macro script for a nested time loop in side a time series
i took a macro generated by the experiment designer that does a time
series of z stack of 2 channels, with points visiting. I added a DO loop
around the points visiting loop with an other TLAPSE command for it to
wait after each tome it does all the points. Have to subtract this
TLAPSE value from the main DO loop TLAPSE as it will have already waited
after then end of doing the last set of points, and so the main loop
time lapse is that wait minus the whole time wait needed.